Build table columns dynamically
