+++ src/dataTypes.js
@@ -1,0 +1,12 @@
+const dataTypes = [
+  { id: 'genetic_association', label: 'Genetic associations (variants)' },
+  { id: 'genetic_literature', label: 'Genetic associations (genes)' },
+  { id: 'somatic_mutation', label: 'Somatic mutations' },
+  { id: 'known_drug', label: 'Drugs' },
+  { id: 'affected_pathway', label: 'Pathways & systems biology' },
+  { id: 'literature', label: 'Text mining' },
+  { id: 'rna_expression', label: 'RNA expression' },
+  { id: 'animal_model', label: 'Animal models' },
+];
+
+export default dataTypes;
